Fix some graphical issues
