+++ src/level_instantiation/spawning/objects/npc.rs
@@ -1,13 +1,11 @@
 use crate::file_system_interaction::asset_loading::AnimationAssets;
+use crate::level_instantiation::spawning::objects::player;
 use crate::level_instantiation::spawning::objects::CollisionLayer;
 use crate::movement::general_movement::{CharacterAnimations, CharacterControllerBundle};
 use crate::movement::navigation::Follower;
 use crate::world_interaction::dialog::DialogTarget;
 use bevy::prelude::*;
 use bevy_xpbd_3d::prelude::*;
-
-pub(crate) const HEIGHT: f32 = 0.4;
-pub(crate) const RADIUS: f32 = 0.4;
 
 pub(crate) fn spawn(
     follower: Query<Entity, Added<Follower>>,
@@ -18,7 +16,7 @@ pub(crate) fn spawn(
         commands
             .entity(entity)
             .insert((
-                CharacterControllerBundle::capsule(HEIGHT, RADIUS),
+                CharacterControllerBundle::capsule(player::HEIGHT, player::RADIUS),
                 Follower,
                 CharacterAnimations {
                     idle: animations.character_idle.clone(),
@@ -33,7 +31,7 @@ pub(crate) fn spawn(
             .with_children(|parent| {
                 parent.spawn((
                     Name::new("NPC Dialog Collider"),
-                    Collider::cylinder(HEIGHT / 2., RADIUS * 5.),
+                    Collider::cylinder(player::HEIGHT / 2., player::RADIUS * 5.),
                     CollisionLayers::new([CollisionLayer::Sensor], [CollisionLayer::Player]),
                     Sensor,
                 ));
